feat: Allow sorting invoices by invoice amount
This commit adds the ability to sort invoices by their total
invoice amount. A mapping for 'invoice_amount' is included in
the $sortMapping array within the index method of the
InvoiceController. This allows users to sort the invoice list
based on the 'invoice_amount' field. The dd() statement was
also commented out.