fix: Update amount calculation and dependencies in CheckRequisitionForm
This commit addresses an issue where the `php_amount` and
`amount_in_words` were not being updated correctly when selecting
invoices. The total amount calculation is now performed directly
within the `useEffect` hook, instead of relying on the `selectedTotal`
useMemo, to ensure up-to-date values. Also, removed `selectedTotal`
from the dependency array of useEffect hook and added `invoices` so
that form is updated when invoices are refetched. Also uses formatted
SINumbers instead of normal SINumbers.